Initialize Client

+++ server/controllers/messageController.js
@@ -1,9 +1,10 @@
 const asyncHandler = require('express-async-handler')
 const Message = require('../models/messageModel')
+const User = require('../models/userModel')
 
 // ROUTE    /messages/
-// PRIV?    Private Access
-// DESC.    @GET (Read) All Messages
+// PRIV?    Public Access
+// DESC.    @GET (Read) All (Public) Messages
 const getAllMessages = asyncHandler(async (req, res) => {
 
     const messages = await Message.find()
@@ -11,8 +12,18 @@ const getAllMessages = asyncHandler(async (req, res) => {
     res.status(200).json(messages)
 })
 
+// ROUTE    /messages/
+// PRIV?    Public Access
+// DESC.    @GET (Read) All (Public) Messages
+const getCurrentUserMessages = asyncHandler(async (req, res) => {
+
+    const messages = await Message.find({ user: req.user.id })
+
+    res.status(200).json(messages)
+})
+
 // ROUTE    /messages/:id
-// PRIV?    Private Access
+// PRIV?    Public Access
 // DESC.    @GET (Read) Single Message
 const getSingleMessage = asyncHandler(async (req, res) => {
     res.json({ message: `Read Single Message ${req.params.id}`})
@@ -29,6 +40,7 @@ const postSingleMessage = asyncHandler(async (req, res) => {
 
     const message = await Message.create({
         text: req.body.text,
+        user: req.user.id
     })
 
     res.status(200).json(message)
@@ -44,6 +56,19 @@ const putSingleMessage = asyncHandler(async (req, res) => {
     if (!message) {
         res.status(400)
         throw new Error('Message not found!')
+    }
+
+    const user = await User.findById(req.user.id)
+
+    if (!user) {
+        res.status(401)
+        throw new Error('User not found!')
+    }
+
+    // Make sure the logged-in user matches the message's user
+    if (message.user.toString() !== user.id) {
+        res.status(401)
+        throw new Error('User not authorized!')
     }
 
     const updatedMessage = await Message.findByIdAndUpdate(req.params.id, req.body, {new: true})
@@ -63,6 +88,19 @@ const deleteSingleMessage = asyncHandler(async (req, res) => {
         throw new Error('Message not found!')
     }
 
+    const user = await User.findById(req.user.id)
+
+    if (!user) {
+        res.status(401)
+        throw new Error('User not found!')
+    }
+
+    // Make sure the logged-in user matches the message's user
+    if (message.user.toString() !== user.id) {
+        res.status(401)
+        throw new Error('User not authorized!')
+    }
+
     const deletedMessage = await Message.findByIdAndDelete(req.params.id, req.body)
 
     res.status(200).json(deletedMessage)
@@ -70,5 +108,5 @@ const deleteSingleMessage = asyncHandler(async (req, res) => {
 
 
 module.exports = {
-    getAllMessages, postSingleMessage, getSingleMessage, putSingleMessage, deleteSingleMessage
+    getAllMessages, getCurrentUserMessages, postSingleMessage, getSingleMessage, putSingleMessage, deleteSingleMessage
 }
